Fix bug in the uniform cylindrical grid
In uniform cylindrical grid, the extended inner boundary for the
DGF calculation would overshoot to R<0 region if too many cells are
added inside the inner boundary. In order to prevent this, the condition
R_min > 16 dR, or equivalently, Rmax/Rmin < 1 + NR/16 should be satisfied,
as described in the last paragraph of the Appendix B in Moon et al. (2019)